Move perf smoke tests to dedicated Tests/PerformanceTests csproj with own workflow
Agent-Logs-Url: https://github.com/gui-cs/Terminal.Gui/sessions/5a2d2af9-446d-4bb7-b955-49055dc0e17d
