+++ Tests/PerformanceTests/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+global using Terminal.Gui.App;
+global using Terminal.Gui.Drivers;
+global using Terminal.Gui.Input;
+global using Terminal.Gui.ViewBase;
+global using Terminal.Gui.Views;
+global using Terminal.Gui.Drawing;
+global using UnitTests;
